fix: resolve all 9 audit findings (4 medium + 5 low)
Medium fixes:
- M1: btc-bridge now uses env vars instead of hardcoded BRIDGE_KEY
- M2: executor CHAIN_ID configurable via AINCORE_CHAIN_ID env var (defaults to TESTNET)
- M3: evm_client threshold field now used in mint_tokens logging
- M4: removed unused BitcoinTx/TxInput/TxOutput structs from btc_client

Low fixes:
- L1: DA verification explicitly wired into consensus block finality
- L2: monitor node list configurable via AINCORE_NODES env var
- L3: governance vote weights upgraded u64 -> u128 to match staking module
- L4: removed debug print from main.rs
- L5: added data/aincore_data/testnet_data to .gitignore

+++ governance/governance/src/lib.rs
@@ -6,7 +6,7 @@ use storage::StateDB;
 
 #[derive(Debug, Serialize, Deserialize)]
 struct AccountData {
-    pub balance: u64,
+    pub balance: u128, // L3 FIX: Match staking module's u128 coin representation
     pub sequence_number: u64,
     pub public_key: String,
 }
@@ -41,8 +41,8 @@ pub struct Proposal {
     pub start_time: u64,
     pub end_time: u64,
     pub execution_time: Option<u64>, // TimeLock: Earliest execution time
-    pub yes_votes: u64,
-    pub no_votes: u64,
+    pub yes_votes: u128, // L3 FIX: Match u128 balance type
+    pub no_votes: u128,
     pub status: ProposalStatus,
 }
 
@@ -60,7 +60,7 @@ pub struct Vote {
     pub proposal_id: String,
     pub voter: String,
     pub approve: bool,
-    pub weight: u64, // Stake amount
+    pub weight: u128, // L3 FIX: Stake amount in u128 to match coin module
 }
 
 pub struct GovernanceManager {
